Move text field from incubator to components folder
