+++ src/testkit/index.ts
@@ -3,7 +3,7 @@ export {ImageDriver} from '../components/image/Image.driver';
 export {TextDriver} from '../components/text/Text.driver';
 export {SwitchDriver} from '../components/switch/switch.driver';
 export {ButtonDriver} from '../components/button/Button.driver';
-export {TextFieldDriver} from '../incubator/TextField/TextField.driver';
+export {TextFieldDriver} from '../components/textField/TextField.driver';
 export {PickerDriver} from '../components/picker/Picker.driver';
 export {CarouselDriver} from '../components/carousel/Carousel.driver';
 export {CheckboxDriver} from '../components/checkbox/Checkbox.driver';
